tools/phase1/requirements.txt
# Phase 1 Orchestrator - Dependencies

# Basis
pyyaml>=6.0.1

# Template-Engine für Handlebars
pybars3>=0.9.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Validierung (später benötigt)
jsonschema>=4.19.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Embeddings (für Synonym-Erkennung, falls gewünscht)
# sentence-transformers>=2.2.0
